cambios varios

+++ app/js/index.js
@@ -9,18 +9,25 @@ function editEntry(id){
 
 async function deleteEntry(id){
 
-    sql = "DELETE FROM horoscopos WHERE id=" + id + ";"; 
-    res = await fetch('/api/update_entry.php', {
-        method: 'POST',
-        body: sql
-    })
-
-    res = await res.text();
-    if (res != 'success') {
-        alert(res);
-    }else{
-        alert("Actualización realizada con éxito");
-        window.location.reload();
+    let confirmation = confirm("¿Deseas eliminar el horóscopo seleccionado?");
+    if (confirmation){
+        sql = "DELETE FROM horoscopos WHERE id=" + id + ";"; 
+        res = await fetch('/api/update_entry.php', {
+            method: 'POST',
+            body: sql
+        })
+    
+        res = await res.text();
+        if (res != 'success') {
+            alert(res);
+        }else{
+            alert("Actualización realizada con éxito");
+            window.location.reload();
+        }
     }
+    else{
+        window.alert("ANULADO: no se ha eliminado el elemento seleccionado");
+    }
+    
 
 }
